Tidy up VS2010 projects as much as it would allow + fix a couple more compiler warnings in 64-bit + time issues (the latter in 64-bit only in fault processing)
git-svn-id: svn://svn.code.sf.net/p/passwordsafe/code/trunk/pwsafe/pwsafe@3818 1f79f812-37fb-46fe-a122-30589dd2bf55

+++ src/ui/Windows/ControlExtns.cpp
@@ -347,7 +347,7 @@ void CEditExtn::OnContextMenu(CWnd* pWnd, CPoint point)
   iter = std::find_if(m_vmenu_items.begin(), m_vmenu_items.end(),
                       equal_cmd(nCmd));
   if (iter != m_vmenu_items.end()) {
-    this->GetParent()->SendMessage(nCmd);
+    this->GetParent()->SendMessage((UINT)nCmd);
     return;
   }
 
@@ -357,7 +357,7 @@ void CEditExtn::OnContextMenu(CWnd* pWnd, CPoint point)
     case WM_COPY:
     case WM_CLEAR:
     case WM_PASTE:
-      SendMessage(nCmd);
+      SendMessage((UINT)nCmd);
       break;
     case EM_SELECTALL:
       SendMessage(EM_SETSEL, 0, -1);
